+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/exceptions/BasketServiceException.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket.exceptions;
+package org.cryptimeleon.incentive.services.basket.exceptions;
 
 public class BasketServiceException extends Exception {
     public BasketServiceException(String message) {
