+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/exceptions/BasketUnauthorizedException.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket.exceptions;
+package org.cryptimeleon.incentive.services.basket.exceptions;
 
 public class BasketUnauthorizedException extends BasketServiceException {
     public BasketUnauthorizedException(String message) {
