+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/exceptions/ItemNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket.exceptions;
+package org.cryptimeleon.incentive.services.basket.exceptions;
 
 public class ItemNotFoundException extends BasketServiceException {
     public ItemNotFoundException() {
