+++ services/basket/src/main/java/org/cryptimeleon/incentive/services/basket/exceptions/WrongBasketValueException.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket.exceptions;
+package org.cryptimeleon.incentive.services.basket.exceptions;
 
 public class WrongBasketValueException extends BasketServiceException {
     public WrongBasketValueException() {
