+++ services/basket/src/test/java/org/cryptimeleon/incentive/services/basket/GenerateSwagger.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket;
+package org.cryptimeleon.incentive.services.basket;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
