+++ services/basket/src/test/java/org/cryptimeleon/incentive/services/basket/RedeemTest.java
@@ -1,4 +1,4 @@
-package org.cryptimeleon.incentive.basket;
+package org.cryptimeleon.incentive.services.basket;
 
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
@@ -12,10 +12,10 @@ import org.springframework.test.web.reactive.server.WebTestClient;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.cryptimeleon.incentive.basket.ClientHelper.*;
+import static org.cryptimeleon.incentive.services.basket.ClientHelper.*;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class PayTest {
+public class RedeemTest {
 
     Logger logger = LoggerFactory.getLogger(BasketController.class);
 
@@ -24,53 +24,6 @@ public class PayTest {
 
     @Value("${basket-service.redeem-secret}")
     private String redeemSecret;
-
-    @Test
-    public void payTest(@Autowired WebTestClient webTestClient) {
-        logger.info("Error message for not existing basket");
-        var wrongBasketId = UUID.randomUUID();
-        payBasket(webTestClient, wrongBasketId, 3, HttpStatus.NOT_FOUND, paymentSecret);
-
-        logger.info("Creating new basket and adding items");
-        var createResponse = createBasket(webTestClient);
-        UUID basketId = createResponse.getResponseBody();
-
-        logger.info("Paying empty basket is not possible");
-        payBasket(webTestClient, basketId, 0, HttpStatus.BAD_REQUEST, paymentSecret);
-
-        var itemsResponse = getItems(webTestClient);
-        var items = itemsResponse.getResponseBody();
-
-        var firstTestItem = items[0];
-        var secondTestItem = items[1];
-
-        putItem(webTestClient, basketId, firstTestItem.getId(), 3, HttpStatus.OK);
-        putItem(webTestClient, basketId, secondTestItem.getId(), 1, HttpStatus.OK);
-
-        var basket = queryBasket(webTestClient, basketId).getResponseBody();
-        assertThat(basket.getItems())
-                .containsEntry(firstTestItem.getId(), 3)
-                .containsEntry(secondTestItem.getId(), 1);
-        assertThat(basket.isPaid()).isFalse();
-
-        logger.info("Paying basket with wrong value not possible");
-        payBasket(webTestClient, basketId, basket.getValue() + 1, HttpStatus.BAD_REQUEST, paymentSecret);
-
-        logger.info("Paying basket with correct parameters works");
-        payBasket(webTestClient, basketId, basket.getValue(), HttpStatus.OK, paymentSecret);
-        basket = queryBasket(webTestClient, basketId).getResponseBody();
-        assertThat(basket.isPaid()).isTrue();
-
-        logger.info("Cannot alter paid basket");
-        putItem(webTestClient, basketId, firstTestItem.getId(), 5, HttpStatus.BAD_REQUEST);
-        deleteBasketItem(webTestClient, basketId, secondTestItem.getId(), HttpStatus.BAD_REQUEST);
-        assertThat(basket.getItems())
-                .containsEntry(firstTestItem.getId(), 3)
-                .containsEntry(secondTestItem.getId(), 1);
-
-        logger.info("Delete basket");
-        deleteBasket(webTestClient, basketId);
-    }
 
     @Test
     void redeemTest(@Autowired WebTestClient webTestClient) {
